Remove database connection from DatabaseInitializer.
As its name suggests DatabaseInitializer is used to initialize DB
parameters of context objects. In particular it is used in the
DatabaseExtension constructor. However such initialization is made
before CI containers gets launched, which make it impossible to
construct DatabaseExtension with correct parameters. As such,
DatabaseInitializer only set DB parameters in the context, leaving
it up to the latest to make the actual connection.